finishing refactor entry point
