+++ redux-intro/package.json
@@ -4,17 +4,20 @@
   "description": "Lessons from redux videos",
   "main": "index.js",
   "scripts": {
-    "build": "npm run test:lint && webpack",
+    "build": "run-s clean test webpack",
+    "clean": "rimraf dist",
     "watch": "webpack --watch",
     "start": "webpack-dev-server --open",
-    "test": "jest",
+    "test": "run-s test:lint test:unit",
     "test:lint": "tslint -p . -c tslint.json",
+    "test:unit": "jest --coverage",
     "watch:test": "jest --watch"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
     "@types/jest": "^22.1.2",
+    "@types/lodash": "^4.14.102",
     "@types/node-uuid": "0.0.28",
     "@types/react": "^16.0.36",
     "@types/react-dom": "^16.0.3",
@@ -22,9 +25,8 @@
     "css-loader": "^0.28.9",
     "html-webpack-plugin": "^2.30.1",
     "jest": "^22.2.2",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "redux": "^3.7.2",
+    "npm-run-all": "^4.1.2",
+    "rimraf": "^2.6.2",
     "style-loader": "^0.20.1",
     "ts-jest": "^22.0.3",
     "ts-loader": "^3.5.0",
@@ -38,23 +40,27 @@
       "^.+\\.tsx?$": "ts-jest"
     },
     "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ]
+    "moduleFileExtensions": [ "ts", "tsx", "js", "jsx", "json", "node" ]
   },
   "prettier": {
     "printWidth": 120,
     "singleQuote": true
   },
+  "collectCoverFrom": [ "**/*{ts,tsx}", "!**/node_modules/**" ],
   "dependencies": {
-    "@types/lodash": "^4.14.102",
     "lodash": "^4.17.5",
     "node-uuid": "^1.4.8",
-    "react-redux": "^5.0.6"
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-redux": "^5.0.6",
+    "redux": "^3.7.2"
+  },
+  "coverageThreshold": {
+    "global": {
+      "branches": 10,
+      "functions": 10,
+      "lines": 10,
+      "statements": 10
+    }
   }
 }
