+++ redux-intro/src/todo/configureStore.ts
@@ -1,0 +1,28 @@
+import { throttle } from 'lodash';
+import { Store } from 'react-redux';
+import { createStore } from 'redux';
+import todoReducer, { TodoAppState } from './reducers';
+import { loadState, saveState } from './util';
+
+const devtoolExtension = (window as any).__REDUX_DEVTOOLS_EXTENSION__;
+
+const configureStore = (): Store<TodoAppState> => {
+  const persistedState = loadState();
+
+  const store = createStore<TodoAppState>(
+    todoReducer,
+    persistedState,
+    devtoolExtension && devtoolExtension(),
+  );
+
+  store.subscribe(
+    throttle(() => {
+      saveState({
+        todos: store.getState().todos,
+      });
+    }, 1000),
+  );
+  return store;
+};
+
+export default configureStore;
