--- redux-intro/src/todo/index.ts
@@ -1,2 +1,0 @@
-export { todos, Todo } from './todoReducer';
-export { visibilityFilter } from './visibilityFilterReducer';
