+++ redux-intro/src/todo/reducers/index.ts
@@ -1,0 +1,15 @@
+import { combineReducers } from 'redux';
+import { Todo, todos } from './todoReducer';
+import { visibilityFilter } from './visibilityFilterReducer';
+
+export interface TodoAppState {
+  todos: Todo[];
+  visibilityFilter: string;
+}
+
+const todoReducer = combineReducers<TodoAppState>({
+  todos,
+  visibilityFilter,
+});
+
+export default todoReducer;
